+++ projects/ngx-forms-inputs/src/inputs/index.ts
@@ -11,3 +11,4 @@ export * from './time-input/time-input.component';
 export * from './search-input/search-input.component';
 export * from './text-input/text-input.component';
 export * from './tag-list-input/tag-list-input.component';
+export * from './date-time-input/date-time-input.component';