+++ projects/ngx-forms-inputs/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @juulsgaard/ngx-forms-inputs
 
+## 16.0.31
+
+### Patch Changes
+
+- Improve date time input parsing
+
+## 16.0.30
+
+### Patch Changes
+
+- Fix DateTime input tag
+
+## 16.0.29
+
+### Patch Changes
+
+- Add DateTime input
+- Updated dependencies
+  - @juulsgaard/ngx-tools@16.1.25
+  - @juulsgaard/ngx-forms@16.0.15
+
 ## 16.0.28
 
 ### Patch Changes